requirements.txt
fastapi
uvicorn
ultralytics
Pillow
numpy
psycopg2
python-multipart
